requirements.txt
Mastodon.py==1.5.1
openai==0.27.0
SQLAlchemy==1.4.28
mysql-connector-python==8.0.28
python-dotenv==0.19.2
requests==2.26.0
cryptography==36.0.1
